back/requirements.txt
fastapi
uvicorn[standard]
httpx
google-generativeai
pydantic
pydantic-settings
qdrant-client
pandas
openpyxl
sentence-transformers
torch
structlog
prometheus-fastapi-instrumentator
slowapi
python-multipart
sentencepiece
tiktoken
rapidfuzz==3.10.0
PyYAML